+++ services/agent/internal/usecases/sender/http/sender.go
@@ -1,4 +1,4 @@
-package sender
+package http
 
 import (
 	"context"
